virtual-trading-sim/requirements.txt
Flask==3.0.3
mysql-connector-python==8.3.0
APScheduler==3.10.4
bcrypt==4.1.2
python-dotenv==1.0.1
